+++ src/utils/__tests__/isStandardValue-test.js
@@ -2,11 +2,14 @@ import test from "tape"
 import isStandardValue from "../isStandardValue"
 
 test("isStandardValue", t => {
-  t.ok(isStandardValue("initial"))
-  t.ok(isStandardValue("currentColor"))
-  t.ok(isStandardValue("10px"))
-  t.ok(isStandardValue("45deg"))
-  t.notOk(isStandardValue("$sass-variable"))
-  t.notOk(isStandardValue("@less-variable"))
+  t.ok(isStandardValue("initial"), "keyword")
+  t.ok(isStandardValue("currentColor"), "svg keyword")
+  t.ok(isStandardValue("10px"), "dimension")
+  t.ok(isStandardValue("45deg"), "angle")
+  t.notOk(isStandardValue("$sass-variable"), "scss var")
+  t.notOk(isStandardValue("@less-variable"), "less var")
+  t.notOk(isStandardValue("#{$var}"), "scss interpolation")
+  t.notOk(isStandardValue("@{var}"), "less interpolation")
+
   t.end()
 })